requirements.txt
streamlit
requests
moralis
python-dotenv
plotly
selenium
beautifulsoup4